requirements.txt
httpx
pydantic-settings
python-dotenv
discord.py
sentry-sdk